Finished game. Decorations added. Custom game story created. New Entety added. Custom behaviour added. Goomba and Koopa images changed.

+++ public/src/ts/traits/PlayerController.ts
@@ -7,12 +7,14 @@ export default class PlayerController extends Trait {
     public checkPoint:Vec2;
     public time:number;
     public score:number;
+    public gameOver:boolean;
 
     constructor() {
         super('playerKontroller');
         this.checkPoint = new Vec2(0,0);
         this.time = 300;
         this.score = 0;
+        this.gameOver = false;
     }
 
     setPlayer(entity: Entity) {
@@ -23,12 +25,22 @@ export default class PlayerController extends Trait {
     }
 
     update(entity: Entity, deltaTime: number, level: Level) {
+        
+        if(this.gameOver){
+            return;
+        }
+
         if (!level.entities.has(this.player)) {
             this.player.killable.revive();
             this.player.pos.set(this.checkPoint.x, this.checkPoint.y);
             level.entities.add(this.player)
         }else{
             this.time -= deltaTime * 2; //TODO kill when time is finished;
+        }
+
+        if(this.time <= 0){
+            this.player.killable.dead = true;
+            this.gameOver = true;
         }
 
         this.player.killable.onScreenHandler(this.player.pos.y);
